refactor アロー関数を関数に直し、d.tsを消す
